@@ -11,17 +11,18 @@
     },
     "devDependencies": {
         "@popperjs/core": "^2.10.2",
+        "@vue/compiler-sfc": "^3.2.31",
         "axios": "^0.21",
         "bootstrap": "^5.1.3",
-        "laravel-mix": "^6.0.6",
+        "laravel-mix": "^6.0.0-beta.17",
         "lodash": "^4.17.19",
         "postcss": "^8.1.14",
         "resolve-url-loader": "^3.1.2",
         "sass": "^1.32.11",
         "sass-loader": "^11.0.1",
-        "vue": "^2.6.12",
-        "vue-loader": "^15.9.8",
-        "vue-template-compiler": "^2.6.12",
-        "bootstrap-vue": "^2.21.2"
+        "vue": "^3.2.31",
+        "vue-loader": "^16.0.0-beta.9",
+        "vue-router": "^4.0.12",
+        "vuex": "^4.0.2"
     }
 }
